+++ CHANGELOG.md
@@ -20,6 +20,9 @@ Smarter handling of keybindings:
   a bug if you miss this).
 * Correctly handle `\<foo-map>` in docstrings.
 
+When `helpful-max-buffers` is set, we now kill buffers in the order of
+least recently accessed. Previously, we killed in creation order.
+
 # v0.5
 
 Allow function tracing to be enabled/disabled from Helpful buffers.